feat(agentctl): rebuild cli with Effect
