+++ services/jangar/agentctl/README.md
@@ -21,11 +21,39 @@ brew install proompteng/tap/agentctl
 
 ## Modes
 - **Kube mode (default):** uses kubeconfig + context and talks to the Kubernetes API directly.
-- **gRPC mode (optional):** uses the Jangar gRPC endpoint.
+- **gRPC mode (optional):** uses the Jangar gRPC endpoint; enable with `--grpc` or `AGENTCTL_MODE=grpc`.
+
+## Quickstart
+
+### Kube mode (default)
+
+```bash
+agentctl status
+agentctl agent list
+agentctl impl list
+```
+
+### gRPC mode
+
+```bash
+kubectl -n agents port-forward svc/agents-grpc 50051:50051
+agentctl --grpc --server 127.0.0.1:50051 status
+```
+
+### Help & discovery
+
+```bash
+agentctl help
+agentctl examples
+agentctl help run
+agentctl agent --help
+```
 
 ## Configuration
 
 The CLI reads configuration from `~/.config/agentctl/config.json` (or `$XDG_CONFIG_HOME/agentctl/config.json`).
+
+`agentctl config view` masks tokens by default; use `--show-secrets` to display them.
 
 Supported fields:
 
@@ -34,6 +62,7 @@ Supported fields:
   "address": "agents-grpc.agents.svc.cluster.local:50051",
   "namespace": "agents",
   "token": "optional-shared-token",
+  "tls": false,
   "kubeconfig": "/path/to/kubeconfig",
   "context": "my-context"
 }
@@ -43,6 +72,7 @@ Environment overrides:
 
 - `AGENTCTL_SERVER` (preferred)
 - `AGENTCTL_ADDRESS`
+- `AGENTCTL_MODE` (`grpc` or `kube`)
 - `AGENTCTL_NAMESPACE`
 - `AGENTCTL_TOKEN`
 - `AGENTCTL_TLS` (`1` to enable TLS)
@@ -55,11 +85,15 @@ Environment overrides:
 ## Usage
 
 ```bash
+agentctl examples
 agentctl version
 agentctl version --client
 agentctl config view
+agentctl config view --show-secrets
 agentctl config set --namespace agents --address 127.0.0.1:50051
+agentctl config set --tls
 agentctl --grpc --server 127.0.0.1:50051 status
+agentctl completion install zsh
 
 agentctl agent list
 agentctl agent get <name>
@@ -70,6 +104,9 @@ agentctl agent delete <name>
 
 agentctl impl list
 agentctl impl create --text "..." --summary "..." --source provider=github,externalId=...,url=...
+agentctl impl create --text @impl.md --summary "..."
+agentctl impl create --text - --summary "..." < impl.md
+agentctl impl init --apply
 agentctl impl describe <name>
 agentctl impl watch
 agentctl impl apply -f impl.yaml
@@ -125,6 +162,8 @@ agentctl workspace describe <name>
 agentctl workspace apply -f workspace.yaml
 
 agentctl run submit --agent <name> --impl <name> --runtime <type>
+agentctl run init --apply --wait
+agentctl run codex --prompt "Summarize repo" --agent <name> --runtime workflow --wait
 agentctl run list
 agentctl run get <name>
 agentctl run describe <name>
@@ -135,16 +174,19 @@ agentctl run logs <name> --follow
 agentctl run cancel <name>
 ```
 
-By default, `agentctl` targets the in-cluster service address `agents-grpc.agents.svc.cluster.local:50051` in the
-`agents` namespace. Use `--namespace`/`-n` to override the namespace and `--server` (or `--address`) for
-port-forwarded access. `--output` supports `table` (default), `json`, and `yaml`; `describe` defaults to `yaml` when
-`--output` is omitted.
+Notes:
+- All `apply` commands accept `-f -` to read manifests from stdin.
+- `impl create --text` accepts inline text, `@file`, or `-` for stdin.
+
+By default, `agentctl` targets the Kubernetes API using your kube context. Use `--grpc` (or `AGENTCTL_MODE=grpc`) plus
+`--server` (or `--address`) to target the Jangar gRPC endpoint. `--output` supports `table` (default), `json`, and
+`yaml`; `describe` defaults to `yaml` when `--output` is omitted.
 
 Port-forward example:
 
 ```bash
 kubectl -n agents port-forward svc/agents-grpc 50051:50051
-agentctl --server 127.0.0.1:50051 status
+agentctl --grpc --server 127.0.0.1:50051 status
 ```
 
 ## Build
